pyproject.toml
[project]
name = "progressive-stable-diffusion"
version = "0.1.0"
description = "Ordinal-aware diffusion research implementation using PyTorch Lightning and Hydra."
requires-python = ">=3.10"
readme = "README.md"
authors = [ { name = "Umut DUNDAR", email = "umutdundar37@gmail.com" } ]
keywords = ["diffusion", "pytorch-lightning", "hydra", "ordinal", "stable-diffusion"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Framework :: PyTorch",
]
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pytorch-lightning>=2.1.0",
    "hydra-core>=1.3.2,<1.4",
    "omegaconf>=2.3,<2.4",
    "torchaudio>=2.1.0",
    "wandb>=0.15",
    "numpy>=1.25",
    "einops>=0.6",
    "accelerate>=0.25",
    "opencv-python-headless>=4.8",
    "pandas>=2.1",
    "diffusers>=0.31.0",
    "Pillow>=10",
]


[project.optional-dependencies]
metrics = [
    "torchmetrics[image]>=1.0",
    "lpips>=0.1.4",
    "scikit-image>=0.21"
]

dev = [
    "black>=24.9",
    "isort>=5.12",
    "flake8>=7.0",
    "mypy>=1.10",
    "ruff>=0.1",
]

[build-system]
requires = ["setuptools>=65.0", "wheel>=0.40"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]
include = "src/"

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.flake8]
max-line-length = 100
exclude = ["__pycache__", "venv", ".venv", "build"]
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = 3.10
warn_unreachable = true
warn_unused_configs = true
ignore_missing_imports = true
files = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "D", "B"]
ignore = ["D106"]
exclude = ["build", "dist", "__pycache__"]
